@@ -23,38 +23,39 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.12.1
+  http: ^0.13.3
   random_string: ^2.1.0
-  expandable_bottom_sheet: ^0.2.1+1
-  expandable: ^4.1.4
-  path_provider: ^1.6.11
-  hive: ^1.4.4+1
-  hive_flutter: ^0.3.1
-  flutter_bloc: ^6.1.2
-  bloc: ^6.1.1
-  google_sign_in: ^4.5.6
-  kiwi: ^2.0.0
-  flutter_native_admob: ^2.1.0
-  eva_icons_flutter: ^2.0.0
-  url_launcher: ^5.5.0
-  flutter_email_sender: ^3.0.1
-  firebase_core: ^0.5.1
-  firebase_messaging: ^6.0.16
-  firebase_auth: ^0.18.2
-  firebase_database: ^4.2.0
-  json_serializable: ^3.5.0
+  expandable_bottom_sheet: ^1.1.0
+  expandable: ^5.0.1
+  path_provider: ^2.0.2
+  hive: ^2.0.4
+  hive_flutter: ^1.1.0
+  flutter_bloc: ^7.0.1
+  bloc: ^7.0.0
+  google_sign_in: ^5.0.4
+  kiwi: ^3.0.0
+  eva_icons_flutter: ^3.0.0
+  url_launcher: ^6.0.9
+  flutter_email_sender: ^5.0.2
+  firebase_core: ^1.4.0
+  firebase_messaging: ^10.0.4
+  firebase_auth: ^3.0.0
+  firebase_database: ^7.1.2
+  json_serializable: ^4.1.4
   flutter_auth_buttons: ^0.10.0
+  google_mobile_ads: ^0.13.2
+
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
+  cupertino_icons: ^1.0.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  hive_generator: ^0.8.2
-  build_runner: ^1.11.1
-  flutter_launcher_icons: "^0.7.3"
-  device_preview: ^0.4.7
+  hive_generator: ^1.1.0
+  build_runner: ^2.0.6
+  flutter_launcher_icons: ^0.9.0
+  device_preview: ^0.7.4
 
 
 flutter_icons:
