@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:google_mobile_ads/google_mobile_ads.dart';
 import 'package:hive/hive.dart';
 import 'package:random_waifu/models/models.dart';
 
@@ -12,14 +13,14 @@ class CollectionStateLoading extends CollectionState {
 
 class CollectionStateLoaded extends CollectionState {
   final Box<dynamic> waifus;
-  const CollectionStateLoaded(this.waifus);
+  final NativeAd ad;
+  const CollectionStateLoaded(this.waifus, this.ad);
 }
 
 class CollectionStateAdLoaded extends CollectionState {
   final adLoaded;
   const CollectionStateAdLoaded(this.adLoaded);
 }
-
 
 class CollectionStateError extends CollectionState {
   final String error;
