@@ -1,56 +1,28 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_native_admob/flutter_native_admob.dart';
-import 'package:flutter_native_admob/native_admob_controller.dart';
+import 'package:google_mobile_ads/google_mobile_ads.dart';
 import 'package:hive/hive.dart';
 import 'package:random_waifu/app_config.dart';
 import 'package:random_waifu/screens/collection/widgets/waifu_list_item.dart';
 
 class CollectionLoaded extends StatefulWidget {
   final Box<dynamic> waifus;
-  CollectionLoaded({Key key, this.waifus});
+  final NativeAd ad;
+  CollectionLoaded({Key key, this.waifus, this.ad});
 
   @override
   _CollectionLoadedState createState() => _CollectionLoadedState();
 }
 
 class _CollectionLoadedState extends State<CollectionLoaded> {
-  final _adUnitID = AppConfig().adId;
-  double adHeight = 0;
-  final NativeAdmobController _controller = NativeAdmobController();
-  StreamSubscription _subscription;
-
   @override
   void initState() {
-    _subscription = _controller.stateChanged.listen(_onStateChanged);
-
     super.initState();
-  }
-
-  void _onStateChanged(AdLoadState state) {
-    switch (state) {
-      case AdLoadState.loading:
-        setState(() {
-          adHeight = 0;
-        });
-        break;
-
-      case AdLoadState.loadCompleted:
-        setState(() {
-          adHeight = 100;
-        });
-        break;
-
-      default:
-        break;
-    }
   }
 
   @override
   void dispose() {
-    _subscription.cancel();
-    _controller.dispose();
     super.dispose();
   }
 
@@ -73,21 +45,19 @@ class _CollectionLoadedState extends State<CollectionLoaded> {
           decoration: BoxDecoration(
             border: Border(
               top: BorderSide(
-                color:
-                    adHeight == 0 ? Colors.transparent : Colors.grey.shade200,
+                color: widget.ad == null
+                    ? Colors.transparent
+                    : Colors.grey.shade200,
               ),
             ),
           ),
           width: double.infinity,
-          height: adHeight,
-          child: NativeAdmob(
-            adUnitID: _adUnitID,
-            controller: _controller,
-            type: NativeAdmobType.full,
-            loading: Center(
-              child: CircularProgressIndicator(),
-            ),
-          ),
+          height: widget.ad != null ? 100 : 0,
+          child: widget.ad != null
+              ? AdWidget(
+                  ad: widget.ad,
+                )
+              : Container(),
         ),
       ],
     );
